+++ gluecodium/src/main/java/com/here/gluecodium/GluecodiumOptions.kt
@@ -26,5 +26,6 @@ data class GluecodiumOptions(
     var commonOutputDir: String = "",
     var generators: Set<String> = setOf(),
     var isValidatingOnly: Boolean = false,
-    var isEnableCaching: Boolean = false
+    var isEnableCaching: Boolean = false,
+    var isStrictMode: Boolean = false
 )