+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gluecodium project Release Notes
 
+## Unreleased
+### Features:
+  * Added `-strict` command line parameter to enable "strict" validation mode for `@Immutable` structs.
+
 ## 10.4.2
 Release date: 2021-12-10
 ### Bug fixes: